+++ CSharp/runtime/DictionaryFromXml.cs
@@ -107,6 +107,9 @@ public class DictionaryFromXml
             case "ellipse":
                 _currentDictionary.Add("ellipse", true);
                 return Error.Ok;
+            case "capsule":
+                _currentDictionary.Add("capsule", true);
+                return Error.Ok;
         }
 
         var dictKey = elementName;