+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.4] - 2025-12-03
+
+### Fixed
+- Fixed issue when using a tileset offset (luckily rarely or never used)
+
+### Added
+- Added support for the new capsule object / collision shape coming with Tiled 1.12
+
 ## [2.2.3] - 2025-11-26
 
 ### Fixed